--- src/ru/rudenko/Parity.java
@@ -1,9 +1,0 @@
-package ru.rudenko;
-
-public class Parity {
-    public static boolean checkParityByBinary(int argument) {
-        String binaryArgument = Integer.toBinaryString(argument);
-        char lastChar = binaryArgument.charAt(binaryArgument.length() - 1);
-        return lastChar == '0';
-    }
-}
